Refactored screen classes to separate modules that are imported to GUI
